backend/pyproject.toml
[project]
name = "youtube-to-slides"
version = "0.1.0"
description = "Web server to generate PowerPoint presentations from YouTube videos"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "youtube-transcript-api>=1.2.4",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "anthropic>=0.40.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
